Some fixes for FreeRTOS

+++ src/drivers/f4-delay/delay.h
@@ -1,4 +1,4 @@
 #pragma once
 #include "stm32f4xx_hal.h"
 
-void delay(int ms);
+void delay(long ms);
